fix(clippy): make the workspace clippy-clean and gate it in CI (#6) (#248)
Completes review finding #6. Adds a `lint` CI job running
`cargo clippy --workspace --all-targets --all-features -- -D warnings`
(plus the fmt check, moved here from the per-OS matrix so it runs once).

Approach for a numerical + pyo3 crate: fix the genuine lints, and deliberately
allow (crate-level, with documented reasons) the ones that are domain-idiomatic,
API-essential, or tool noise:
- needless_range_loop (~184): index loops are the clearer idiom for the matrix
  math; rewriting risks bugs in hot numerical paths.
- too_many_arguments / type_complexity: the fit/binding entry points are
  inherently many-parameter (the option-struct refactor is tracked separately).
- upper_case_acronyms: the public model types (LDA, STM, DMR, ...) — renaming
  breaks the API.
- useless_conversion (~188): emitted inside pyo3 0.22's #[pymethods] codegen,
  not our code.
- doc_lazy_continuation / doc_overindented_list_items: rustdoc cosmetics (the
  user docs are mkdocs).

Genuinely fixed: manual slice copies -> copy_from_slice (detm, sts); map_identity
(cluster); duplicated #[allow] (hdp); sort_by -> sort_by_key (output);
`!(x > 0.0)` -> explicit `x <= 0.0 || x.is_nan()` (lightlda, python); is_some+
unwrap -> if-let (prodlda load); plus the machine-applicable `clippy --fix` pass.
False positives annotated locally with reasons (misnamed_getters on the
`learned_`-prefixed HDP getters; mistyped_literal_suffixes on the nmf magic hex;
field_reassign on the external HDbscan struct). The bins carry the same
domain-lint allow header (they are separate crate roots).

Verified: cargo clippy -D warnings clean, cargo fmt --check clean, cargo test
--workspace (133 + 16), maturin --features python, pytest (2782 passed). No
behavior change; faSTM surface untouched.

+++ src/gsdmm.rs
@@ -152,7 +152,7 @@ impl GsdmmModel {
                             continue;
                         }
                         let base = self.nw[kk][w] as f64 + self.beta;
-                        for j in 0..cw as u32 {
+                        for j in 0..cw {
                             lp += (base + j as f64).ln();
                         }
                     }
@@ -234,7 +234,7 @@ fn resample_doc<R: Rng>(model: &mut GsdmmModel, d: usize, doc: &[u32], rng: &mut
                 continue;
             }
             let base = model.nw[kk][w] as f64 + model.beta;
-            for j in 0..cw as u32 {
+            for j in 0..cw {
                 lp += (base + j as f64).ln();
             }
         }
